HTMLStripper/requirements.txt
beautifulsoup4==4.12.0
docx==0.2.4
pandas==2.1.1
fpdf==1.7.2
ebooklib==0.18
language-tool-python==1.5.4
pyspellchecker==0.6.2
tqdm==4.66.1
tkinter==8.6
